Rename package coop.rchain.models.rholang.{sort => sorter}
The `sort` name disturbed use of such-named methods within rholang
package.
