--- models/src/main/scala/coop/rchain/models/rholang/sort/package.scala
@@ -1,3 +1,0 @@
-package coop.rchain.models.rholang
-
-package object sort extends ScoreTree
