+++ models/src/main/scala/coop/rchain/models/rholang/sorter/BoolSortMatcher.scala
@@ -1,10 +1,10 @@
-package coop.rchain.models.rholang.sort
+package coop.rchain.models.rholang.sorter
 
 import cats.effect.Sync
 import coop.rchain.models.Expr.ExprInstance.GBool
 import cats.implicits._
 
-private[sort] object BoolSortMatcher extends Sortable[GBool] {
+private[sorter] object BoolSortMatcher extends Sortable[GBool] {
   def sortMatch[F[_]: Sync](g: GBool): F[ScoredTerm[GBool]] =
     if (g.value) {
       ScoredTerm(g, Leaves(Score.BOOL, 0)).pure[F]
