+++ models/src/main/scala/coop/rchain/models/rholang/sorter/package.scala
@@ -1,0 +1,3 @@
+package coop.rchain.models.rholang
+
+package object sorter extends ScoreTree
