+++ models/src/test/scala/coop/rchain/models/SortedParHashSetSpec.scala
@@ -9,7 +9,7 @@ import coop.rchain.crypto.hash.Blake2b256
 import coop.rchain.models.Expr.ExprInstance._
 import coop.rchain.models.Var.VarInstance.BoundVar
 import coop.rchain.models.rholang.implicits._
-import coop.rchain.models.rholang.sort.ordering._
+import coop.rchain.models.rholang.sorter.ordering._
 import org.scalatest.{Assertion, FlatSpec, Matchers}
 
 import scala.collection.immutable.BitSet
