+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/Substitute.scala
@@ -9,7 +9,7 @@ import coop.rchain.models.Expr.ExprInstance._
 import coop.rchain.models.Var.VarInstance._
 import coop.rchain.models._
 import coop.rchain.models.rholang.implicits._
-import coop.rchain.models.rholang.sort._
+import coop.rchain.models.rholang.sorter._
 import coop.rchain.rholang.interpreter.accounting.{Chargeable, Cost, CostAccountingAlg}
 import coop.rchain.rholang.interpreter.errors.SubstituteError
