+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/SortSpec.scala
@@ -2,9 +2,9 @@ package coop.rchain.rholang.interpreter
 import coop.rchain.models.Var.VarInstance.FreeVar
 import coop.rchain.models._
 import coop.rchain.models.rholang.implicits._
-import coop.rchain.models.rholang.sort.Sortable
+import coop.rchain.models.rholang.sorter.Sortable
 import org.scalatest.{FlatSpec, Matchers}
-import coop.rchain.models.rholang.sort.ScoredTerm
+import coop.rchain.models.rholang.sorter.ScoredTerm
 import monix.eval.Coeval
 
 import scala.collection.immutable.BitSet
